requirements.txt
# Flask
# Flask-PyMongo
# pymongo
# tensorflow
# scikit-learn
# numpy
# Werkzeug
# gunicorn
# python-dotenv

Flask==3.0.3
Flask-PyMongo==2.3.0
pymongo==4.6.3
tensorflow-cpu==2.13.0
scikit-learn==1.3.2
numpy==1.24.3
Werkzeug==3.0.3
gunicorn==21.2.0
python-dotenv==1.0.1

